+++ infrastructure-kotlin/src/main/kotlin/kotlin/internal/Kotlin.kt
@@ -4,18 +4,14 @@ import com.redmadrobot.build.InfrastructurePlugin
 import com.redmadrobot.build.dsl.isRunningOnCi
 import com.redmadrobot.build.internal.InternalGradleInfrastructureApi
 import com.redmadrobot.build.internal.findBooleanProperty
-import org.gradle.api.JavaVersion
 import org.gradle.api.Project
-import org.gradle.api.provider.Provider
 import org.gradle.kotlin.dsl.*
-import org.jetbrains.kotlin.gradle.dsl.JvmTarget
 
 @InternalGradleInfrastructureApi
-public fun InfrastructurePlugin.configureKotlin(jvmTargetProperty: Provider<JavaVersion>) {
+public fun InfrastructurePlugin.configureKotlin() {
     val warningsAsErrors = project.getWarningsAsErrorsProperty()
     project.kotlinCompile {
         compilerOptions {
-            jvmTarget = JvmTarget.fromTarget(jvmTargetProperty.get().toString())
             allWarningsAsErrors = warningsAsErrors
             freeCompilerArgs.add("-opt-in=kotlin.RequiresOptIn")
         }